Extended coverage of unit tests, added spotbugs.
1. All Consume worker checks for url extensions of files to omit.
2. Consume workers now dispatch produce workers using the executor pool.
3. After processing consumers check if cancel request submitted.
4. After cancel request submitted no new jobs are enqueued.
5. Modified the order of launch parameters 2nd parameter is now timeout.

+++ src/main/java/com/acabra/jwebcrawler/model/CrawledNode.java
@@ -25,11 +25,11 @@ public class CrawledNode {
         this.parentId = parentId;
     }
 
-    public CrawledNode buildChild(String link, int nextId) {
+    public CrawledNode buildChild(String link, long nextId) {
         return new CrawledNode(link, nextId, this.level + 1, this.id);
     }
 
-    public CrawledNode redirection(String redirectURI, int nextId) {
+    public CrawledNode redirection(String redirectURI, long nextId) {
         return new CrawledNode(redirectURI, nextId, this.level, this.parentId);
     }
